requirements.txt
streamlit==1.32.0
supabase==2.3.4
plotly==5.19.0
pandas==2.2.1
python-dotenv==1.0.1
websockets>=14.0
